pumpdetection_model/requirements.txt
# requirements.txt
adal==1.2.7
applicationinsights==0.11.10
arch==5.3.1
argcomplete==2.1.2
asttokens==2.2.1
attrs==23.1.0
azure-common==1.1.28
azure-core==1.27.1
azure-graphrbac==0.61.1
azure-identity==1.13.0
azure-mgmt-authorization==3.0.0
azure-mgmt-containerregistry==10.1.0
azure-mgmt-core==1.4.0
azure-mgmt-keyvault==10.2.2
azure-mgmt-resource==22.0.0
azure-mgmt-storage==21.0.0
azure-storage-blob==12.13.0
azure-storage-queue==12.6.0
azureml-automl-core==1.52.0.post1
azureml-automl-dnn-vision==1.52.0.post1
azureml-automl-runtime==1.52.0.post1
azureml-core==1.52.0
azureml-dataprep==4.11.4
azureml-dataprep-native==38.0.0
azureml-dataprep-rslex==2.18.4
azureml-dataset-runtime==1.52.0
azureml-defaults==1.52.0
azureml-inference-server-http==0.8.4
azureml-interpret==1.52.0
azureml-mlflow==1.52.0
azureml-responsibleai==1.52.0
azureml-telemetry==1.52.0
azureml-train-automl-client==1.52.0
azureml-train-automl-runtime==1.52.0
azureml-train-core==1.52.0
azureml-train-restclients-hyperdrive==1.52.0
azureml-training-tabular==1.52.0
backcall==0.2.0
backports-tempfile==1.0
backports-weakref==1.0.post1
bcrypt==4.0.1
bokeh==2.4.3
boto==2.49.0
cachetools==5.3.1
captum==0.5.0
cffi==1.15.1
colorcet==3.0.1
contextlib2==21.6.0
cryptography==41.0.0
cython==0.29.14
dask==2023.2.0
databricks-cli==0.17.7
dataclasses==0.6
debugpy==1.6.7
decorator==5.1.1
dice-ml==0.9
dill==0.3.6
distributed==2023.2.0
distro==1.8.0
docker==6.1.3
dotnetcore2==3.1.23
econml==0.14.1
entrypoints==0.4
erroranalysis==0.4.4
executing==1.2.0
fairlearn==0.8.0
fire==0.5.0
flask==2.2.5
flask-cors==3.0.10
flatbuffers==23.5.26
fsspec==2023.6.0
fusepy==3.0.1
fvcore==0.1.5.post20221221
gensim==3.8.3
gitdb==4.0.10
gitpython==3.1.31
google-api-core==2.11.1
google-auth==2.21.0
googleapis-common-protos==1.59.1
gunicorn==20.1.0
h5py==3.9.0
humanfriendly==10.0
imageio==2.31.1
importlib-metadata==6.8.0
inference-schema==1.5.1
interpret-community==0.29.0
interpret-core==0.3.2
iopath==0.1.10
ipykernel==6.8.0
ipython==8.12.2
isodate==0.6.1
itsdangerous==2.1.2
jedi==0.18.2
jeepney==0.8.0
jinja2==3.1.2
jmespath==0.10.0
jsonpickle==3.0.1
jsonschema==4.18.0
jsonschema-specifications==2023.6.1
jupyter-client==7.4.9
jupyter-core==5.3.1
keras2onnx==1.6.0
knack==0.10.1
lightgbm==3.2.1
locket==1.0.0
markupsafe==2.1.2
matplotlib==3.6.3
matplotlib-inline==0.1.6
ml-wrappers==0.4.11
mlflow-skinny==2.4.1
msal==1.22.0
msal-extensions==1.0.0
msgpack==1.0.5
msrest==0.7.1
msrestazure==0.6.4
munch==4.0.0
ndg-httpsclient==0.5.1
nest-asyncio==1.5.6
networkx==2.5
nose==1.3.7
numba==0.53.1
oauthlib==3.2.2
onnx==1.13.1
onnxconverter-common==1.6.0
onnxmltools==1.4.1
onnxruntime==1.11.1
opencensus==0.11.2
opencensus-context==0.1.3
opencensus-ext-azure==1.1.9
opencv-python-headless==4.3.0.38
packaging==23.0
param==1.13.0
paramiko==3.2.0
parso==0.8.3
partd==1.4.0
pathspec==0.11.1
patsy==0.5.3
pexpect==4.8.0
pickleshare==0.7.5
pillow==9.0.1
pkginfo==1.9.6
pkgutil-resolve-name==1.3.10
pmdarima==1.8.0
portalocker==2.7.0
pretrainedmodels==0.7.4
prompt-toolkit==3.0.39
property-cached==1.6.4
protobuf==3.20.3
ptyprocess==0.7.0
pure-eval==0.2.2
pyarrow==11.0.0
pyasn1==0.5.0
pyasn1-modules==0.3.0
pycparser==2.21
pyct==0.5.0
pydantic==1.10.11
pygments==2.15.1
pyjwt==2.7.0
pynacl==1.5.0
pynvml==8.0.4
pyopenssl==23.2.0
pytorch-ignite==0.4.12
pywavelets==1.4.1
pyyaml==6.0
pyzmq==25.1.0
raiutils==0.4.0
referencing==0.29.1
requests-oauthlib==1.3.1
resnest==0.0.6b20210504
responsibleai==0.28.0
rpds-py==0.8.8
rsa==4.9
s3transfer==0.5.2
saliency==0.1.3
scikit-image==0.19.3
scipy==1.5.3
secretstorage==3.3.3
semver==2.13.0
simplification==0.5.7
skl2onnx==1.4.9
sklearn-pandas==1.7.0
smart-open==1.9.0
smmap==5.0.0
sortedcontainers==2.4.0
sparse==0.14.0
sqlparse==0.4.4
stack-data==0.6.2
statsmodels==0.11.1
tabulate==0.9.0
tblib==2.0.0
termcolor==2.3.0
tifffile==2023.7.4
timm==0.4.12
toolz==0.12.0
torch==1.11.0
torchvision==0.12.0
tornado==6.3.2
traitlets==5.9.0
urllib3==1.26.16
wcwidth==0.2.6
websocket-client==1.6.1
werkzeug==2.3.6
wrapt==1.12.1
yacs==0.1.8
zict==3.0.0
boto3==1.20.19
botocore==1.23.19
brotli==1.0.9
brotli-bin==1.0.9
brotli-python==1.0.9
ca-certificates==2023.5.7
certifi==2023.5.7
charset-normalizer==3.1.0
cloudpickle==2.2.1
colorama==0.4.6
contourpy==1.1.0
cycler==0.11.0
fonttools==4.40.0
freetype==2.12.1
idna==3.4
importlib-resources==5.12.0
importlib_resources==5.12.0
joblib==1.2.0
kiwisolver==1.4.4
lcms2==2.15
lerc==4.0.0
libblas==3.9.0
libbrotlicommon==1.0.9
libbrotlidec==1.0.9
libbrotlienc==1.0.9
libcblas==3.9.0
libdeflate==1.18
libffi==3.3
libgfortran-ng==13.1.0
libgfortran5==13.1.0
libjpeg-turbo==2.1.5.1
liblapack==3.9.0
libllvm10==10.0.1
libopenblas==0.3.23
libpng==1.6.39
libsqlite==3.42.0
libtiff==4.5.1
libwebp-base==1.3.1
libxcb==1.15
libzlib==1.2.13
llvmlite==0.36.0
munkres==1.1.4
numpy==1.22.3
openjpeg==2.5.0
pandas==1.1.5
pip==21.3.1
platformdirs==3.8.1
pooch==1.7.0
psutil==5.8.0
pthread-stubs==0.4
pycocotools==2.0.4
pyparsing==3.0.9
pysocks==1.7.1
python-dateutil==2.8.2
python-tzdata==2023.3
python_abi==3.8
pytz==2023.3
recordclass==0.14.3
requests==2.31.0
scikit-learn==0.22.2.post1
setuptools==65.5.1
six==1.16.0
slicer==0.0.7
tbb==2020.2
threadpoolctl==3.1.0
tqdm==4.65.0
typing-extensions==4.7.1
typing_extensions==4.7.1
unicodedata2==15.0.0
wheel==0.38.1
xorg-libxau==1.0.11
xorg-libxdmcp==1.1.3
zipp==3.15.0
zstd==1.5.2